[misc] sync with changes of logger in ppl.common

+++ src/ppl/nn/engines/arm/engine.cc
@@ -166,7 +166,7 @@ ArmEngine::ConfHandlerFunc ArmEngine::conf_handlers_[] = {};
 
 RetCode ArmEngine::Configure(uint32_t option, ...) {
     if (option >= ENGINE_CONF_MAX) {
-        LOG(ERROR) << "invalid option[" << option << "] >= [" << ENGINE_CONF_MAX << "]";
+        LOG(ERROR) << "invalid option[" << option << "] >= [" << (uint32_t)ENGINE_CONF_MAX << "]";
         return RC_INVALID_VALUE;
     }
